requirements.txt
# ============================================
# ASSISTENTE JURÍDICO IA - DEPENDÊNCIAS
# ============================================

# --- LangChain Core ---
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13

# --- OpenAI ---
openai>=1.10.0

# --- Processamento de PDF ---
pdfplumber>=0.10.3
PyPDF2>=3.0.1

# --- Processamento de Documentos ---
python-docx>=1.1.0
openpyxl>=3.1.2

# --- Manipulação de Dados ---
pandas>=2.1.4
numpy>=1.26.2

# --- Interface Web ---
streamlit>=1.29.0

# --- Banco de Dados Vetorial (RAG) ---
chromadb>=0.4.22

# --- Variáveis de Ambiente ---
python-dotenv>=1.0.0

# --- Validação de Dados ---
pydantic>=2.5.3

# --- Requisições HTTP ---
requests>=2.31.0

# --- Utilitários ---
tiktoken>=0.5.2
tenacity>=8.2.3

# --- Embeddings ---
sentence-transformers>=2.2.2

# --- OCR para PDFs escaneados ---
# pytesseract>=0.3.10
# Pillow>=10.1.0